# RESTapi/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laraval Challenge",
        "description": "Swagger documentation for the Laraval Challenge API",
        "version": "1.0.0"
    },
    "paths": {
        "/api/v1/debit-cards": {
            "get": {
                "tags": [
                    "Bank Cards"
                ],
                "summary": "Get all debit cards for the authenticated user",
                "description": "Returns all debit cards owned by the authenticated user, including card type details.",
                "operationId": "da095ae83d03bdc769aec33c442f7feb",
                "responses": {
                    "200": {
                        "description": "List of bank cards",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 201
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 2
                                                    },
                                                    "user_id": {
                                                        "type": "integer",
                                                        "example": 2
                                                    },
                                                    "bank_card_type_id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "bank_card_number": {
                                                        "type": "integer",
                                                        "example": 1234567898765432
                                                    },
                                                    "cvv": {
                                                        "type": "integer",
                                                        "example": 321
                                                    },
                                                    "expiry_date": {
                                                        "type": "string",
                                                        "format": "date",
                                                        "example": "2029-03-24"
                                                    },
                                                    "bank_card_name": {
                                                        "type": "string",
                                                        "example": "Test Debit Card"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2025-03-24T12:07:09.000000Z"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2025-03-24T12:07:09.000000Z"
                                                    },
                                                    "deleted_at": {
                                                        "type": "string",
                                                        "format": "nullable",
                                                        "example": null
                                                    },
                                                    "bank_card_type": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer",
                                                                "example": 1
                                                            },
                                                            "name": {
                                                                "type": "string",
                                                                "example": "Debit Card"
                                                            },
                                                            "slug": {
                                                                "type": "string",
                                                                "example": "debit-card"
                                                            },
                                                            "created_at": {
                                                                "type": "string",
                                                                "format": "date-time",
                                                                "example": "2025-03-24T12:05:02.000000Z"
                                                            },
                                                            "updated_at": {
                                                                "type": "string",
                                                                "format": "date-time",
                                                                "example": "2025-03-24T12:05:02.000000Z"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Bank Cards"
                ],
                "summary": "Create a new debit card",
                "description": "Creates a new debit card for the authenticated user. The card number, CVV, and expiry date are generated automatically.",
                "operationId": "9b6cd7cfca18579ef45282521c24ae4d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "bank_card_name"
                                ],
                                "properties": {
                                    "bank_card_name": {
                                        "type": "string",
                                        "example": "My New bank Card"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Card created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 201
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "My New bank Card successfully created"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "501": {
                        "description": "Failed to create card",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 501
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "My New bank Card could not be created"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}